fix for del file

+++ examples/api-samples/inc_samples/sample18.php
@@ -137,8 +137,9 @@
         if (count($item) > 0) {
             for ($i = 0; $i < count($item); $i++) {
                 $next = $path . "\\" . $item[$i];
-
+                if (file_exists($next)) {
                     unlink($next);
+                }
             }
         }
         rmdir($path);
